Extract Resolve
